backend-python/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==2.3.1
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==42.0.7
cycler==0.12.1
distro==1.9.0
dlib==19.24.4
docker-pycreds==0.4.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.9.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.43.1
frozenlist==1.4.1
fsspec==2023.9.2
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.38
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.18.0
idna==3.4
importlib-metadata==7.0.0
importlib-resources==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.5.0
mdurl==0.1.2
ml-dtypes==0.3.2
mmcv-full==1.4.2
mmdet==2.20.0
mpmath==1.3.0
multidict==6.0.5
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.0
numpy==1.26.4
opencv-python==3.4.18.65
opt-einsum==3.3.0
optree==0.11.0
oss2==2.18.5
packaging==23.2
pandas==2.1.1
pathtools==0.1.2
Pillow==9.3.0
platformdirs==4.1.0
protobuf==4.24.4
psutil==5.9.6
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytorch-msssim==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
rich==13.7.1
safetensors==0.4.0
scikit-build==0.17.6
scikit-learn==1.3.2
scipy==1.11.3
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools-scm==8.0.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spark-ai-python==0.3.24
sympy==1.12
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-logger==0.1.0
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.2.0
timm==0.6.5
tomli==2.0.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.13
wandb==0.15.12
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
ws4py==0.5.1
wsproto==1.2.0
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
